Prevents duplicate FITID
When transactions with the same amount occured on the same day a duplicate FITID was created. This gave issues when loading the data. Fixed the issue by appending time and accountto to the FITID. In some special cases this could still result in a duplicate FITID (sometimes time is not specified and accountto could be the same). This is fixed by checking whether the ID already exists and if so add an incremental number to ensure uniqueness.